requirements.txt
python==3.9.19
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.1.4
matplotlib==3.9.2
plotly==5.24.1
streamlit==1.38.0
streamlit_option_menu==0.3.13
scikit-surprise==1.5.2